refactor(index.js): Extract the nested switch inside the for loop. This is an example of nested dependency first step to solve. The way it was before,any change in the for loop outside or even in the invoice play variables above could break the switch code fragment.
According to the book, to extract it we need to check if the variables present in the code fragment will lose its reference declaration.

I opted, initially to declare "thisAmount" out of amountFor method and just pass it down to this extracted method. This is a little divergence from the book. However, it can be declared inside amountFor method according to your preference.

My first conclusion here is: We must start from the most nested code fragment ever. I need to confirm it later.